+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - GP2-2472 - Increase padding on example component
 - GP2-2473 - Increase padding on lesson component
 - GP2-2471 - Update Tooltip component
+- GP2-2464 / GP2-2480 - improve button text baseline
 - GP2-2491 - Increase spacing at bottom of forms
 
 ### Fixed bugs